Added support for following types or namespaces:
- System.Threading
- System.Threading.Tasks
- Windows.System.Threading (UWP threading)
- Windows.Devices.I2c
- Windows.Devices.Spi

Also:
+ Support for Context switch for M3 and M4 (w and w/o floating point, no lazy stacking yet)
+ Robustified SDK (still needs plenty of work)
+ Revised LLVM translation
+ Revised and refined HW model
+ Laid foundations for ref counting GC

fixes #25
fixes #17
fixes #12
fixes #11
fixes #9

+++ VisualStudio/LlilumProjectType/LlilumApplication/LlilumApplication.ProjectType/LlilumDebuggerLaunchProvider.cs
@@ -49,15 +49,19 @@ namespace LlilumApplication
             return !string.IsNullOrEmpty(commandValue);
         }
 
-        public string CreateDebuggerFile(string workingDirectory, string executablePath, string gdbPath, string gdbArgs, string script)
+        public string CreateDebuggerFileIfNoneExist(string workingDirectory, string executablePath, string gdbPath, string gdbArgs, string script)
         {
             string debuggerFile = string.Format(DebuggerFileFormat, workingDirectory, "tmp_debug");
             
-            using (System.IO.StreamWriter file =
-                new System.IO.StreamWriter(debuggerFile))
+            if(!File.Exists(debuggerFile))
             {
-                file.WriteLine(string.Format(script, gdbPath, executablePath, gdbArgs));
+                using (System.IO.StreamWriter file =
+                new System.IO.StreamWriter(debuggerFile))
+                {
+                    file.WriteLine(string.Format(script, gdbPath, executablePath, gdbArgs));
+                }
             }
+            
             return debuggerFile;
         }
 
@@ -83,7 +87,7 @@ namespace LlilumApplication
             }
 
             // Create the temporary file for passing to the debug engine
-            string debuggerFile = CreateDebuggerFile(dir, executable, gdbPath, gdbArgs, debugScript);
+            string debuggerFile = CreateDebuggerFileIfNoneExist(dir, executable, gdbPath, gdbArgs, debugScript);
 
             settings.Options = string.Format(DebuggerOptionsFormat, debuggerFile);
             settings.LaunchOperation = DebugLaunchOperation.CreateProcess;
@@ -95,6 +99,9 @@ namespace LlilumApplication
 
             if(!string.IsNullOrEmpty(pyOcdPath))
             {
+                // Even though we did it in deploy, do it here just in case we go straight to debug
+                LlilumHelpers.TryKillPyocd();
+
                 ProcessStartInfo start = new ProcessStartInfo();
                 start.FileName = pyOcdPath;
                 start.Arguments = pyOcdArgs;
